debt - move jsonediting/configediting to common
