+++ src/tsconfig.strictNullChecks.json
@@ -285,10 +285,10 @@
 		"./vs/workbench/services/backup/node/backupFileService.ts",
 		"./vs/workbench/services/backup/test/electron-browser/backupFileService.test.ts",
 		"./vs/workbench/services/broadcast/electron-browser/broadcastService.ts",
+		"./vs/workbench/services/configuration/common/configurationEditingService.ts",
+		"./vs/workbench/services/configuration/common/jsonEditingService.ts",
 		"./vs/workbench/services/configuration/node/configuration.ts",
-		"./vs/workbench/services/configuration/node/configurationEditingService.ts",
 		"./vs/workbench/services/configuration/node/configurationService.ts",
-		"./vs/workbench/services/configuration/node/jsonEditingService.ts",
 		"./vs/workbench/services/configuration/test/common/configurationModels.test.ts",
 		"./vs/workbench/services/configuration/test/electron-browser/configurationEditingService.test.ts",
 		"./vs/workbench/services/configuration/test/electron-browser/configurationService.test.ts",
