+++ src/vs/workbench/workbench.main.ts
@@ -108,7 +108,7 @@ import 'vs/workbench/services/editor/browser/codeEditorService';
 import 'vs/workbench/services/broadcast/electron-browser/broadcastService';
 import 'vs/workbench/services/preferences/browser/preferencesService';
 import 'vs/workbench/services/output/node/outputChannelModelService';
-import 'vs/workbench/services/configuration/node/jsonEditingService';
+import 'vs/workbench/services/configuration/common/jsonEditingService';
 import 'vs/workbench/services/textmodelResolver/common/textModelResolverService';
 import 'vs/workbench/services/textfile/common/textFileService';
 import 'vs/workbench/services/dialogs/browser/fileDialogService';
